dailybuild: 軽微な warning の抑止, solution file の一部修正
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7215 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Framework/FWScene.h
@@ -208,7 +208,7 @@ public:
 		return sensor->Cast();
 	}
 	void UpdateSkeletonSensors() {
-		for (int i=0; i<skeletonSensors.size(); ++i) {
+		for (int i=0; i<(int)skeletonSensors.size(); ++i) {
 			skeletonSensors[i]->Update();
 		}
 	}
